[BUGFIX] Handle customer deletion errors
Previously, it was assumed that a customer deletion could fail by means of a falsy return value of the method `Mage_Core_Model_Abstract::delete`; this is not possible since this method either returns `$this` or raises an exception.

In this changeset, fail by falsy return value has been removed, and error handling is added for the case an exception is raised.

Also, the `isSecureArea` registry setting has been cleared right after the deletion operation has succeed or failed.